Split `participants-helper` classes into files
