+++ server/services/feature-flags.ts
@@ -1,6 +1,5 @@
 const strToBoolean = (value) => value === 'true';
-
-module.exports = {
+export default {
   FEATURE_PHASE_ALLOCATION: strToBoolean(process.env.FEATURE_PHASE_ALLOCATION),
   FEATURE_EMPLOYER_FORM: false,
 };
