+++ server/services/participants-helper/PaginatedParticipantsFinder.ts
@@ -1,0 +1,13 @@
+import { run } from './core';
+
+export class PaginatedParticipantsFinder {
+  context;
+
+  constructor(context) {
+    this.context = context;
+  }
+
+  async run() {
+    return run(this.context);
+  }
+}
